refactor(desktop): regroup Exposure UI into Colour + Tone collapsibles
Reorganize the Exposure sidebar for clarity:
- Split the single Exposure panel into two independent top-level collapsibles,
  Colour (white balance + cast removal) and Tone (density/grade/curve/paper),
  as ColourSidebar and ToneSidebar. Each has its own expand state, edited-count
  badge and scoped reset (a section reset no longer clobbers the other's edits).
- Colour header carries a new RGB mini-histogram; Tone keeps the luma one.
- White Balance: region selector is now an icon column (globe/moon/sun) left of
  the CMY sliders, tinting edited-yellow per region; a dynamic
  "Global/Shadows/Highlights White Balance" sub-label; Pick WB is an eyedropper
  icon on that header row.
- Move the shared _icon_toggle/_labeled_toggle helpers to BaseSidebar.
- Default cast_removal_strength set to 0.5.

Pure view-layer change; update keyboard shortcuts, main window, tutorial steps
and their tests to the new sidebars.
