+++ negpy/desktop/view/keyboard_shortcuts.py
@@ -43,7 +43,7 @@ class ShortcutManager:
             "flip_h": lambda: toolbar.flip("horizontal"),
             "flip_v": lambda: toolbar.flip("vertical"),
             "lock_bounds_toggle": lambda: controls.process_sidebar.lock_bounds_btn.toggle(),
-            "pick_wb": lambda: controls.exposure_sidebar.pick_wb_btn.toggle(),
+            "pick_wb": lambda: controls.colour_sidebar.pick_wb_btn.toggle(),
             "manual_crop": lambda: controls.geometry_sidebar.manual_crop_btn.toggle(),
             "auto_crop": lambda: controls.geometry_sidebar.reset_crop_btn.toggle(),
             "pick_dust": lambda: controls.retouch_sidebar.pick_dust_btn.toggle(),
@@ -72,25 +72,25 @@ class ShortcutManager:
         }
 
         slider_targets: dict[str, tuple[Callable[[], object], float]] = {
-            "cyan_inc": (lambda: controls.exposure_sidebar.cyan_slider, 1.0),
-            "cyan_dec": (lambda: controls.exposure_sidebar.cyan_slider, -1.0),
-            "magenta_up": (lambda: controls.exposure_sidebar.magenta_slider, 1.0),
-            "magenta_down": (lambda: controls.exposure_sidebar.magenta_slider, -1.0),
-            "yellow_up": (lambda: controls.exposure_sidebar.yellow_slider, 1.0),
-            "yellow_down": (lambda: controls.exposure_sidebar.yellow_slider, -1.0),
-            "density_up": (lambda: controls.exposure_sidebar.density_slider, 1.0),
-            "density_down": (lambda: controls.exposure_sidebar.density_slider, -1.0),
+            "cyan_inc": (lambda: controls.colour_sidebar.cyan_slider, 1.0),
+            "cyan_dec": (lambda: controls.colour_sidebar.cyan_slider, -1.0),
+            "magenta_up": (lambda: controls.colour_sidebar.magenta_slider, 1.0),
+            "magenta_down": (lambda: controls.colour_sidebar.magenta_slider, -1.0),
+            "yellow_up": (lambda: controls.colour_sidebar.yellow_slider, 1.0),
+            "yellow_down": (lambda: controls.colour_sidebar.yellow_slider, -1.0),
+            "density_up": (lambda: controls.tone_sidebar.density_slider, 1.0),
+            "density_down": (lambda: controls.tone_sidebar.density_slider, -1.0),
             # ISO R scale is inverted: harder grade = lower R.
-            "grade_up": (lambda: controls.exposure_sidebar.grade_slider, -10.0),
-            "grade_down": (lambda: controls.exposure_sidebar.grade_slider, 10.0),
-            "toe_inc": (lambda: controls.exposure_sidebar.toe_slider, 1.0),
-            "toe_dec": (lambda: controls.exposure_sidebar.toe_slider, -1.0),
-            "toe_width_inc": (lambda: controls.exposure_sidebar.toe_w_slider, 1.0),
-            "toe_width_dec": (lambda: controls.exposure_sidebar.toe_w_slider, -1.0),
-            "shoulder_inc": (lambda: controls.exposure_sidebar.sh_slider, 1.0),
-            "shoulder_dec": (lambda: controls.exposure_sidebar.sh_slider, -1.0),
-            "shoulder_width_inc": (lambda: controls.exposure_sidebar.sh_w_slider, 1.0),
-            "shoulder_width_dec": (lambda: controls.exposure_sidebar.sh_w_slider, -1.0),
+            "grade_up": (lambda: controls.tone_sidebar.grade_slider, -10.0),
+            "grade_down": (lambda: controls.tone_sidebar.grade_slider, 10.0),
+            "toe_inc": (lambda: controls.tone_sidebar.toe_slider, 1.0),
+            "toe_dec": (lambda: controls.tone_sidebar.toe_slider, -1.0),
+            "toe_width_inc": (lambda: controls.tone_sidebar.toe_w_slider, 1.0),
+            "toe_width_dec": (lambda: controls.tone_sidebar.toe_w_slider, -1.0),
+            "shoulder_inc": (lambda: controls.tone_sidebar.sh_slider, 1.0),
+            "shoulder_dec": (lambda: controls.tone_sidebar.sh_slider, -1.0),
+            "shoulder_width_inc": (lambda: controls.tone_sidebar.sh_w_slider, 1.0),
+            "shoulder_width_dec": (lambda: controls.tone_sidebar.sh_w_slider, -1.0),
             "offset_inc": (lambda: controls.geometry_sidebar.offset_slider, 1.0),
             "offset_dec": (lambda: controls.geometry_sidebar.offset_slider, -1.0),
             "fine_rot_inc": (lambda: controls.geometry_sidebar.fine_rot_slider, 1.0),
