+++ negpy/desktop/view/main_window.py
@@ -476,7 +476,7 @@ class MainWindow(QMainWindow):
         self.canvas.set_tool_mode(mode)
 
         # We access buttons through the controls panel
-        self.controls_panel.exposure_sidebar.pick_wb_btn.setChecked(mode == ToolMode.WB_PICK)
+        self.controls_panel.colour_sidebar.pick_wb_btn.setChecked(mode == ToolMode.WB_PICK)
         self.controls_panel.geometry_sidebar.manual_crop_btn.setChecked(mode == ToolMode.CROP_MANUAL)
         self.controls_panel.retouch_sidebar.pick_dust_btn.setChecked(mode == ToolMode.DUST_PICK)
