+++ tests/test_tutorial_overlay.py
@@ -13,7 +13,8 @@ _VALID_SECTION_ATTRS = {
     "flatfield_section",
     "geometry_section",
     "process_section",
-    "exposure_section",
+    "colour_section",
+    "tone_section",
     "lab_section",
     "toning_section",
     "retouch_section",
